refactor(node): modularize configuration and runtime lifecycle
Split configuration credentials, network profiles, source-layer merging,
resolved values, resolution, and runtime inputs behind the existing public
facade. Consolidate last-set-value-wins merging without treating explicit
false, zero, or empty values as absent.

Separate startup, RPC composition, and ordered service teardown. Remove
the retired test-only DNS/outbound implementation and unused P2P listener
handle collection; P2pService remains the networking runtime owner. Use
Auth::to_rpc_auth instead of maintaining a duplicate conversion.

Preserve lifecycle/checkpoint regressions with isolated shared fixtures,
add five configuration/rollback/idempotence regressions, and document the
module boundaries. No Cargo features, dependencies, or storage formats
are changed. Native build, Clippy, and test validation remain required.
